+++ detectors/node/opentelemetry-resource-detector-aws/test/detectors/AwsBeanstalkDetector.test.ts
@@ -16,7 +16,7 @@
 
 import * as assert from 'assert';
 import * as sinon from 'sinon';
-import { awsBeanstalkDetector, AwsBeanstalkDetector } from '../../src';
+import { awsBeanstalkDetector, AwsBeanstalkDetectorSync } from '../../src';
 import {
   assertEmptyResource,
   assertServiceResource,
@@ -45,14 +45,15 @@ describe('BeanstalkResourceDetector', () => {
 
   it('should successfully return resource data', async () => {
     fileStub = sinon
-      .stub(AwsBeanstalkDetector, 'fileAccessAsync' as any)
+      .stub(AwsBeanstalkDetectorSync, 'fileAccessAsync' as any)
       .resolves();
     readStub = sinon
-      .stub(AwsBeanstalkDetector, 'readFileAsync' as any)
+      .stub(AwsBeanstalkDetectorSync, 'readFileAsync' as any)
       .resolves(JSON.stringify(data));
     sinon.stub(JSON, 'parse').returns(data);
 
     const resource = await awsBeanstalkDetector.detect();
+    await resource.waitForAsyncAttributes?.();
 
     sinon.assert.calledOnce(fileStub);
     sinon.assert.calledOnce(readStub);
@@ -67,14 +68,15 @@ describe('BeanstalkResourceDetector', () => {
 
   it('should successfully return resource data with noise', async () => {
     fileStub = sinon
-      .stub(AwsBeanstalkDetector, 'fileAccessAsync' as any)
+      .stub(AwsBeanstalkDetectorSync, 'fileAccessAsync' as any)
       .resolves();
     readStub = sinon
-      .stub(AwsBeanstalkDetector, 'readFileAsync' as any)
+      .stub(AwsBeanstalkDetectorSync, 'readFileAsync' as any)
       .resolves(JSON.stringify(noisyData));
     sinon.stub(JSON, 'parse').returns(noisyData);
 
     const resource = await awsBeanstalkDetector.detect();
+    await resource.waitForAsyncAttributes?.();
 
     sinon.assert.calledOnce(fileStub);
     sinon.assert.calledOnce(readStub);
@@ -89,13 +91,14 @@ describe('BeanstalkResourceDetector', () => {
 
   it('should return empty resource when failing to read file', async () => {
     fileStub = sinon
-      .stub(AwsBeanstalkDetector, 'fileAccessAsync' as any)
+      .stub(AwsBeanstalkDetectorSync, 'fileAccessAsync' as any)
       .resolves();
     readStub = sinon
-      .stub(AwsBeanstalkDetector, 'readFileAsync' as any)
+      .stub(AwsBeanstalkDetectorSync, 'readFileAsync' as any)
       .rejects(err);
 
     const resource = await awsBeanstalkDetector.detect();
+    await resource.waitForAsyncAttributes?.();
 
     sinon.assert.calledOnce(fileStub);
     sinon.assert.calledOnce(readStub);
@@ -105,13 +108,14 @@ describe('BeanstalkResourceDetector', () => {
 
   it('should return empty resource when config file does not exist', async () => {
     fileStub = sinon
-      .stub(AwsBeanstalkDetector, 'fileAccessAsync' as any)
+      .stub(AwsBeanstalkDetectorSync, 'fileAccessAsync' as any)
       .rejects(err);
     readStub = sinon
-      .stub(AwsBeanstalkDetector, 'readFileAsync' as any)
+      .stub(AwsBeanstalkDetectorSync, 'readFileAsync' as any)
       .resolves(JSON.stringify(data));
 
     const resource = await awsBeanstalkDetector.detect();
+    await resource.waitForAsyncAttributes?.();
 
     sinon.assert.calledOnce(fileStub);
     sinon.assert.notCalled(readStub);