+++ plugins/node/instrumentation-undici/test/fetch.test.ts
@@ -142,6 +142,9 @@ describe('UndiciInstrumentation `fetch` tests', function () {
         requestHook: () => {
           throw new Error('requestHook error');
         },
+        responseHook: () => {
+          throw new Error('responseHook error');
+        },
         startSpanHook: () => {
           throw new Error('startSpanHook error');
         },
@@ -213,6 +216,12 @@ describe('UndiciInstrumentation `fetch` tests', function () {
             req.headers.push('x-requested-with', 'undici');
           }
         },
+        responseHook: (span, { response }) => {
+          span.setAttribute(
+            'test.response-hook.attribute',
+            response.statusText
+          );
+        },
         startSpanHook: request => {
           return {
             'test.hook.attribute': 'hook-value',
@@ -280,6 +289,11 @@ describe('UndiciInstrumentation `fetch` tests', function () {
         span.attributes['test.hook.attribute'],
         'hook-value',
         'startSpanHook is called'
+      );
+      assert.strictEqual(
+        span.attributes['test.response-hook.attribute'],
+        'OK',
+        'responseHook is called'
       );
     });