+++ metapackages/auto-instrumentations-node/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.49.1](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/auto-instrumentations-node-v0.49.0...auto-instrumentations-node-v0.49.1) (2024-08-05)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @opentelemetry/instrumentation-amqplib bumped from ^0.40.0 to ^0.41.0
+    * @opentelemetry/resource-detector-alibaba-cloud bumped from ^0.28.10 to ^0.29.0
+
 ## [0.49.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/auto-instrumentations-node-v0.48.0...auto-instrumentations-node-v0.49.0) (2024-07-23)